Revert "Supabase Auth #12"
