--- .env.local.example
@@ -1,3 +1,0 @@
-NEXT_PUBLIC_SUPABASE_URL=<your_supabase_project_url>
-NEXT_PUBLIC_SUPABASE_ANON_KEY=<your_supabase_anon_key>
-NEXT_PUBLIC_REDIRECT_URL=http://localhost:3000 # or the production URL
