+++ .gitignore
@@ -2,4 +2,3 @@ node_modules/
 .next/
 .swc/
 .*.env
-.env*.local
