--- app/(auth)/actions.ts
@@ -1,86 +1,0 @@
-"use server";
-
-import { revalidatePath } from "next/cache";
-import { redirect } from "next/navigation";
-
-import { createClient } from "@/utils/supabase/server";
-import { SignUpWithPasswordCredentials } from "@supabase/supabase-js";
-
-const supabase = createClient();
-
-export async function signin(formData: FormData) {
-
-  // type-casting here for convenience
-  // in practice, you should validate your inputs
-  const data = {
-    email: formData.get("email") as string,
-    password: formData.get("password") as string,
-  };
-
-  const { error } = await supabase.auth.signInWithPassword(data);
-
-  if (error) {
-    return {error: error.message };
-  }
-
-  revalidatePath("/", "layout");
-  redirect("/");
-}
-
-export async function signup(formData: FormData) {
-  // type-casting here for convenience
-  // in practice, you should validate your inputs
-  const data: SignUpWithPasswordCredentials = {
-    email: formData.get("email") as string,
-    password: formData.get("password") as string,
-    options: {
-      data: {
-        full_name: formData.get("full-name") as string,
-        company_name: formData.get("company-name") as string,
-      }
-    }
-  };
-
-  const { error } = await supabase.auth.signUp(data);
-  
-  if (error) {
-    return {error: error.message };
-  }
-
-  revalidatePath("/", "layout");
-  redirect("/");
-}
-
-// Google OAuth sign-in
-export async function signinWithGoogle() {
-
-  const { data, error } = await supabase.auth.signInWithOAuth({
-    provider: 'google',
-    options: {
-      redirectTo: `${process.env.NEXT_PUBLIC_REDIRECT_URL}/auth/callback`,
-    }
-  });
-  if (error) {
-    redirect("/error");
-  }
-
-  if (data.url) {
-    redirect(data.url) 
-  }
-
-  revalidatePath("/", "layout");
-  redirect("/");
-}
-
-// Reset password
-export async function resetPassword(formData: FormData) {
-
-  const { data, error } = await supabase.auth.resetPasswordForEmail(formData.get("email") as string);
-
-  if (error) {
-    return {error: error.message };
-  }
-
-  revalidatePath("/", "layout");
-  redirect("/");
-}
