--- app/(auth)/reset-password/layout.tsx
@@ -1,7 +1,0 @@
-import Page from "./page";
-export const metadata = {
-  title: 'Reset Password - Open PRO',
-  description: 'Page description',
-}
-
-export default Page;
