--- app/(auth)/signin/layout.tsx
@@ -1,7 +1,0 @@
-import Page from "./page";
-export const metadata = {
-  title: 'Sign In - Open PRO',
-  description: 'Page description',
-}
-
-export default Page;
