--- app/(auth)/signup/layout.tsx
@@ -1,7 +1,0 @@
-import Page from "./page";
-export const metadata = {
-  title: 'Sign Up - Open PRO',
-  description: 'Page description',
-}
-
-export default Page;
