+++ components/ui/header.tsx
@@ -1,6 +1,5 @@
-import Link from "next/link";
-import MobileMenu from "./mobile-menu";
-import AuthButton from "./authbutton";
+import Link from 'next/link'
+import MobileMenu from './mobile-menu'
 import PearLightLogo from './PearLight70x70.svg'
 
 export default function Header() {
@@ -42,10 +41,11 @@ export default function Header() {
               </li>
             </ul>
           </nav>
-          {process.env.NODE_ENV !== 'production' && <AuthButton />} {/* AuthButton is hidden in production */} 
+
           <MobileMenu />
+
         </div>
       </div>
     </header>
-  );
+  )
 }
